Move items into better folders
