+++ src/module/item/base-item.js
@@ -4,7 +4,7 @@ import { itemExtensions } from "./itemExtensions.js";
  * Extend the basic Item with some very simple modifications.
  * @extends {Item}
  */
-export class TorchbearerItem extends Item {
+export class TorchbearerBaseItem extends Item {
   /**
    * Augment the basic Item data model with additional dynamic data.
    */
