+++ src/module/item/index.js
@@ -1,0 +1,3 @@
+export { TorchbearerBaseItem } from "./base-item";
+export { TorchbearerItemSheet } from "./item";
+export { TorchbearerSpellSheet } from "./spell";
