+++ src/module/item/item/index.js
@@ -1,0 +1,1 @@
+export { TorchbearerItemSheet } from './sheet';
