+++ src/module/item/spell/index.js
@@ -1,0 +1,1 @@
+export { TorchbearerSpellSheet } from "./sheet";
