+++ src/module/item/spell/sheet.js
@@ -1,4 +1,4 @@
-import { TorchbearerBaseItemSheet } from "./base-item-sheet.js";
+import { TorchbearerBaseItemSheet } from "../base-item-sheet";
 
 export class TorchbearerSpellSheet extends TorchbearerBaseItemSheet {
   /** @override */
